requirements.txt
Django>=4.2
psycopg2-binary
Pillow
django-crispy-forms
crispy-bootstrap5
